fix: 배포 오류 해결
