+++ src/app/api/actions/userAction.ts
@@ -36,8 +36,6 @@ export async function signup(formData: FormData): Promise<ApiResWithValidation<S
     }
     const fileData: MultiItem<FileRes> = await fileRes.json();
 
-    console.log(fileData);
-
     userObj.image = fileData.item[0].path;
   }
 
@@ -65,6 +63,7 @@ export async function editUser(formData: FormData): Promise<ApiResWithValidation
     address: formData.get('address'),
     image: '',
   };
+  console.log('🚀 ~ editUser ~ userObj:', userObj);
 
   const attach = formData.get('attach') as File;
